Move UI components to UI folder, segregate images
