+++ apps/site/src/components/PageLanding/images/intro/bg-intro-mobile.png
