+++ apps/site/src/components/PageLanding/images/partners/bg-partners.png
