+++ apps/site/src/components/PageLanding/images/signup/bg-signup-discord.png
