+++ apps/site/src/components/PageLanding/images/signup/bg-signup-mobile.png
