+++ apps/site/src/components/PageLanding/json/communities.json
@@ -1,90 +1,90 @@
 [
   {
-    "key": 1,
+    "id": 1,
     "name": "NBA Topshot",
     "image": "/communities/card-topshot1.png",
     "url": "https://nbatopshot.com/"
   },
   {
-    "key": 2,
+    "id": 2,
     "name": "UFC Strike",
     "image": "/communities/card-ufcstrike1.png",
     "url": "https://www.ufcstrike.com/"
   },
   {
-    "key": 3,
+    "id": 3,
     "name": "NBA Topshot",
     "image": "/communities/card-topshot2.png",
     "url": "https://nbatopshot.com/"
   },
   {
-    "key": 4,
+    "id": 4,
     "name": "UFC Strike",
     "image": "/communities/card-ufcstrike2.png",
     "url": "https://www.ufcstrike.com/"
   },
   {
-    "key": 5,
+    "id": 5,
     "name": "Dr Seuss",
     "image": "/communities/card-drseuss.png",
     "url": "https://seussibles.com/"
   },
   {
-    "key": 6,
+    "id": 6,
     "name": "NBA Topshot",
     "image": "/communities/card-topshot1.png",
     "url": "https://nbatopshot.com/"
   },
   {
-    "key": 7,
+    "id": 7,
     "name": "UFC Strike",
     "image": "/communities/card-ufcstrike1.png",
     "url": "https://www.ufcstrike.com/"
   },
   {
-    "key": 8,
+    "id": 8,
     "name": "NBA Topshot",
     "image": "/communities/card-topshot2.png",
     "url": "https://nbatopshot.com/"
   },
   {
-    "key": 9,
+    "id": 9,
     "name": "UFC Strike",
     "image": "/communities/card-ufcstrike2.png",
     "url": "https://www.ufcstrike.com/"
   },
   {
-    "key": 10,
+    "id": 10,
     "name": "Dr Seuss",
     "image": "/communities/card-drseuss.png",
     "url": "https://seussibles.com/"
   },
   {
-    "key": 11,
+    "id": 11,
     "name": "NBA Topshot",
     "image": "/communities/card-topshot1.png",
     "url": "https://nbatopshot.com/"
   },
   {
-    "key": 12,
+    "id": 12,
     "name": "UFC Strike",
     "image": "/communities/card-ufcstrike1.png",
     "url": "https://www.ufcstrike.com/"
   },
   {
-    "key": 13,
+    "id": 13,
     "name": "NBA Topshot",
     "image": "/communities/card-topshot2.png",
     "url": "https://nbatopshot.com/"
   },
   {
-    "key": 14,
+    "id": 14,
     "name": "UFC Strike",
     "image": "/communities/card-ufcstrike2.png",
     "url": "https://www.ufcstrike.com/"
   },
   {
-    "key": 15,
+    "id": 15,
     "name": "Dr Seuss",
     "image": "/communities/card-drseuss.png",
     "url": "https://seussibles.com/"
