+++ apps/site/src/components/UI/index.ts
@@ -1,0 +1,6 @@
+export { default as Button } from './Button';
+export { default as Chip } from './Chip';
+export { default as Footer } from './Footer';
+export { default as Navbar } from './Navbar';
+export { default as NavButtons } from './NavButtons';
+export { default as TextInput } from './TextInput';
